+++ crates/scouter/src/core/mod.rs
@@ -1,6 +1,6 @@
-pub mod alert;
+pub mod cron;
 pub mod error;
-pub mod monitor;
-pub mod num_profiler;
+pub mod profile;
+pub mod spc;
 pub mod stats;
-pub mod string_profiler;
+pub mod utils;
