+++ crates/scouter/src/lib.rs
@@ -1,3 +1,1 @@
 pub mod core;
-pub mod queue;
-pub mod utils;
