--- crates/scouter/src/queue/mod.rs
@@ -1,1 +1,0 @@
-pub mod feature_queue;
