--- crates/scouter/src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod cron;
-pub mod types;
